Cleaned up my poor copying skills, removed rule as its own resource
