+++ pkg/apicontracts/apiresourcecontracts/resource_models_methods_generated.go
@@ -1531,76 +1531,6 @@ func (m ResourceListNetworkpolicies) GetByLabels(search map[string]string) []Res
 	return Response
 }
 
-// Function to return FirewallRule resource by name.
-func (m ResourceListFirewallrules) GetByName(search string) ResourceFirewallRule {
-	for _, resource := range m.Firewallrules {
-		if resource.Metadata.Name == search {
-			return resource
-		}
-	}
-	var emptyResponse ResourceFirewallRule
-	return emptyResponse
-}
-
-// Function to return FirewallRule resource by uid.
-func (m ResourceListFirewallrules) GetByUid(search string) ResourceFirewallRule {
-	for _, res := range m.Firewallrules {
-		if res.Metadata.Uid == search {
-			return res
-		}
-	}
-	var emptyResponse ResourceFirewallRule
-	return emptyResponse
-}
-
-// Function to return FirewallRule resource by label.
-func (m ResourceListFirewallrules) GetByLabels(search map[string]string) []ResourceFirewallRule {
-	var Response []ResourceFirewallRule
-	for _, res := range m.Firewallrules {
-		if len(res.Metadata.Labels) != 0 {
-			if stringhelper.CompareLabels(search, res.Metadata.Labels) {
-				Response = append(Response, res)
-			}
-		}
-	}
-	return Response
-}
-
-// Function to return VirtualMachine resource by name.
-func (m ResourceListVirtualmachines) GetByName(search string) ResourceVirtualMachine {
-	for _, resource := range m.Virtualmachines {
-		if resource.Metadata.Name == search {
-			return resource
-		}
-	}
-	var emptyResponse ResourceVirtualMachine
-	return emptyResponse
-}
-
-// Function to return VirtualMachine resource by uid.
-func (m ResourceListVirtualmachines) GetByUid(search string) ResourceVirtualMachine {
-	for _, res := range m.Virtualmachines {
-		if res.Metadata.Uid == search {
-			return res
-		}
-	}
-	var emptyResponse ResourceVirtualMachine
-	return emptyResponse
-}
-
-// Function to return VirtualMachine resource by label.
-func (m ResourceListVirtualmachines) GetByLabels(search map[string]string) []ResourceVirtualMachine {
-	var Response []ResourceVirtualMachine
-	for _, res := range m.Virtualmachines {
-		if len(res.Metadata.Labels) != 0 {
-			if stringhelper.CompareLabels(search, res.Metadata.Labels) {
-				Response = append(Response, res)
-			}
-		}
-	}
-	return Response
-}
-
 // Function to return FirewallPolicy resource by name.
 func (m ResourceListFirewallpolicies) GetByName(search string) ResourceFirewallPolicy {
 	for _, resource := range m.Firewallpolicies {
