+++ pkg/rorresources/rortypes/resource_interfaces.go
@@ -225,18 +225,6 @@ type NetworkPolicyinterface interface {
 	Get() *ResourceNetworkPolicy
 }
 
-// FirewallRuleinterface represents the interface for resources of the type firewallrule
-type FirewallRuleinterface interface {
-	CommonResourceInterface
-	Get() *ResourceFirewallRule
-}
-
-// VirtualMachineinterface represents the interface for resources of the type virtualmachine
-type VirtualMachineinterface interface {
-	CommonResourceInterface
-	Get() *ResourceVirtualMachine
-}
-
 // FirewallPolicyinterface represents the interface for resources of the type firewallpolicy
 type FirewallPolicyinterface interface {
 	CommonResourceInterface
